refactor: Test enabling cleanups for code generator. (#1242)
