+++ tools/serverpod_cli/lib/src/generator/code_generator.dart
@@ -1,12 +1,7 @@
-import 'dart:io';
-
-import 'package:meta/meta.dart';
+import 'package:code_builder/code_builder.dart';
+import 'package:dart_style/dart_style.dart';
 import 'package:serverpod_cli/analyzer.dart';
-import 'package:serverpod_cli/src/generator/code_generation_collector.dart';
-import 'package:serverpod_cli/src/generator/dart/code_generator_dart.dart';
-import 'package:serverpod_cli/src/generator/psql/legacy_pgsql_generator.dart';
 import 'package:serverpod_cli/src/logger/logger.dart';
-import 'package:serverpod_cli/src/util/internal_error.dart';
 
 /// A code generator is responsible for generating the code for the target
 /// language.
@@ -22,7 +17,6 @@ abstract class CodeGenerator {
   /// Relative paths start at the server package directory.
   ///
   /// Called and generated before [generateProtocolCode].
-  @protected
   Map<String, String> generateSerializableEntitiesCode({
     required List<SerializableEntityDefinition> entities,
     required GeneratorConfig config,
@@ -38,140 +32,29 @@ abstract class CodeGenerator {
   ///
   /// At the time this is called, [generateSerializableEntitiesCode] should
   /// already be called and generated.
-  @protected
   Map<String, String> generateProtocolCode({
     required ProtocolDefinition protocolDefinition,
     required GeneratorConfig config,
   });
-
-  /// List all the directories, that may contain files, that should be cleaned.
-  /// For most [CodeGenerator]s, the output should just depend on [config].
-  ///
-  /// Relative paths start at the server package directory.
-  @protected
-  Future<List<String>> getDirectoriesRequiringCleaning({
-    required ProtocolDefinition protocolDefinition,
-    required GeneratorConfig config,
-  });
-
-  /// The file extensions, this generator uses when generating files.
-  @protected
-  List<String> get outputFileExtensions;
-
-  /// The generators, that run on [generateAll].
-  static const generators = [DartCodeGenerator(), LegacyPgsqlCodeGenerator()];
-
-  /// Generate from [CodeGenerator.generateSerializableEntitiesCode] for all
-  /// [CodeGenerator]s and save the files.
-  ///
-  /// Returns a list of generated files.
-  static Future<List<String>> generateSerializableEntities({
-    required List<SerializableEntityDefinition> entities,
-    required GeneratorConfig config,
-    required CodeGenerationCollector collector,
-  }) async {
-    collector.generatedFiles.clear();
-    var allFiles = {
-      for (var generator in generators)
-        ...generator.generateSerializableEntitiesCode(
-          entities: entities,
-          config: config,
-        )
-    };
-    for (var file in allFiles.entries) {
-      try {
-        log.debug('Generating ${file.key}.');
-        var out = File(file.key);
-        await out.create(recursive: true);
-        await out.writeAsString(file.value, flush: true);
-
-        collector.addGeneratedFile(out);
-      } catch (e, stackTrace) {
-        log.error('Failed to generate ${file.key}!');
-        printInternalError(e, stackTrace);
-      }
-    }
-
-    return allFiles.keys.toList();
-  }
-
-  /// Generate from [CodeGenerator.generateProtocolCode] for all
-  /// [CodeGenerator]s and save the files.
-  ///
-  /// Returns a list of generated files.
-  static Future<List<String>> generateProtocolDefinition({
-    required ProtocolDefinition protocolDefinition,
-    required GeneratorConfig config,
-    required CodeGenerationCollector collector,
-  }) async {
-    collector.generatedFiles.clear();
-    var allFiles = {
-      for (var generator in generators)
-        ...generator.generateProtocolCode(
-          protocolDefinition: protocolDefinition,
-          config: config,
-        )
-    };
-    for (var file in allFiles.entries) {
-      try {
-        log.debug('Generating ${file.key}.');
-        var out = File(file.key);
-        await out.create(recursive: true);
-        await out.writeAsString(file.value, flush: true);
-
-        collector.addGeneratedFile(out);
-      } catch (e, stackTrace) {
-        log.error('Failed to generate ${file.key}');
-        printInternalError(e, stackTrace);
-      }
-    }
-
-    return allFiles.keys.toList();
-  }
-
-  /// Removes files from previous generation runs.
-  /// By removing old files that are not part of the [generatedFiles] in the
-  /// [CodeGenerator.getDirectoriesRequiringCleaning] for each [CodeGenerator].
-  static Future<void> cleanPreviouslyGeneratedFiles({
-    required Set<String> generatedFiles,
-    required ProtocolDefinition protocolDefinition,
-    required GeneratorConfig config,
-  }) async {
-    log.debug('Cleaning up old files.');
-    for (var generator in generators) {
-      var dirs = await generator.getDirectoriesRequiringCleaning(
-          protocolDefinition: protocolDefinition, config: config);
-
-      for (var dir in dirs) {
-        await _removeOldFilesInPath(
-          dir,
-          generatedFiles,
-          generator.outputFileExtensions,
-        );
-      }
-    }
-  }
 }
 
-Future<void> _removeOldFilesInPath(
-  String directoryPath,
-  Set<String> keepPaths,
-  List<String> fileExtensions,
-) async {
-  var directory = Directory(directoryPath);
-  log.debug('Remove old files from $directory');
-  var fileList = await directory.list(recursive: true).toList();
+extension GenerateCode on Library {
+  String generateCode() {
+    var code = accept(DartEmitter.scoped(useNullSafetySyntax: true)).toString();
+    try {
+      return DartFormatter().format('''
+/* AUTOMATICALLY GENERATED CODE DO NOT MODIFY */
+/*   To generate run: "serverpod generate"    */
 
-  for (var entity in fileList) {
-    // Only check Dart files.
-    if (entity is! File ||
-        !fileExtensions.any((extension) => entity.path.endsWith(extension))) {
-      continue;
-    }
+// ignore_for_file: library_private_types_in_public_api
+// ignore_for_file: public_member_api_docs
+// ignore_for_file: implementation_imports
 
-    if (!keepPaths.contains(entity.path)) {
-      log.debug('Remove: $entity');
-      await entity.delete();
+$code
+''');
+    } on FormatterException catch (e) {
+      log.error(e.toString());
     }
+    return code;
   }
 }
