+++ tools/serverpod_cli/test/generator/dart/server_generator.dart/class_field_test.dart
@@ -1,4 +1,4 @@
-import 'package:serverpod_cli/src/generator/dart/code_generator_dart.dart';
+import 'package:serverpod_cli/src/generator/dart/server_code_generator.dart';
 import 'package:test/test.dart';
 import 'package:path/path.dart' as path;
 
@@ -7,7 +7,7 @@ import 'package:serverpod_cli/src/test_util/builders/generator_config_builder.da
 
 const projectName = 'example_project';
 final config = GeneratorConfigBuilder().withName(projectName).build();
-const generator = DartCodeGenerator();
+const generator = DartServerCodeGenerator();
 
 void main() {
   group('Given a class with a none nullable field', () {
@@ -30,19 +30,6 @@ void main() {
 
       expect(codeMap[expectedFileName], contains('String title;'));
     });
-
-    test('then a client-side class is generated with that class variable.', () {
-      var expectedFileName = path.join(
-        '..',
-        'example_project_client',
-        'lib',
-        'src',
-        'protocol',
-        'example.dart',
-      );
-
-      expect(codeMap[expectedFileName], contains('String title;'));
-    });
   });
 
   group('Given a class with a nullable field', () {
@@ -62,19 +49,6 @@ void main() {
     test('then a server-side class is generated with that class variable.', () {
       var expectedFileName =
           path.join('lib', 'src', 'generated', 'example.dart');
-
-      expect(codeMap[expectedFileName], contains('String? title;'));
-    });
-
-    test('then a client-side class is generated with that class variable.', () {
-      var expectedFileName = path.join(
-        '..',
-        'example_project_client',
-        'lib',
-        'src',
-        'protocol',
-        'example.dart',
-      );
 
       expect(codeMap[expectedFileName], contains('String? title;'));
     });
