+++ tools/serverpod_cli/test/generator/dart/server_generator.dart/expected_files_created_test.dart
@@ -1,4 +1,4 @@
-import 'package:serverpod_cli/src/generator/dart/code_generator_dart.dart';
+import 'package:serverpod_cli/src/generator/dart/server_code_generator.dart';
 import 'package:test/test.dart';
 import 'package:path/path.dart' as path;
 
@@ -7,7 +7,7 @@ import 'package:serverpod_cli/src/test_util/builders/generator_config_builder.da
 
 const projectName = 'example_project';
 final config = GeneratorConfigBuilder().withName(projectName).build();
-const generator = DartCodeGenerator();
+const generator = DartServerCodeGenerator();
 
 void main() {
   group('Given a single class when generating the code', () {
@@ -23,34 +23,11 @@ void main() {
       config: config,
     );
 
-    test('then the protocol file is created', () {
-      expect(
-        codeMap.keys,
-        contains(path.join('lib', 'src', 'generated', 'protocol.dart')),
-        reason: 'Expected protocol file to be present, found none.',
-      );
-    });
-
     test('then the server-side file is created', () {
       expect(
         codeMap.keys,
         contains(path.join('lib', 'src', 'generated', 'example.dart')),
         reason: 'Expected server-side file to be present, found none.',
-      );
-    });
-
-    test('then the client-side file is created', () {
-      expect(
-        codeMap.keys,
-        contains(path.join(
-          '..',
-          'example_project_client',
-          'lib',
-          'src',
-          'protocol',
-          'example.dart',
-        )),
-        reason: 'Expected client-side file to be present, found none.',
       );
     });
   });
@@ -72,14 +49,6 @@ void main() {
       config: config,
     );
 
-    test('then the protocol file is created', () {
-      expect(
-        codeMap.keys,
-        contains(path.join('lib', 'src', 'generated', 'protocol.dart')),
-        reason: 'Expected protocol file to be present, found none.',
-      );
-    });
-
     test('then the server-side files are created', () {
       expect(
         codeMap.keys,
@@ -93,38 +62,10 @@ void main() {
         reason: 'Expected server-side file to be present, found none.',
       );
     });
-
-    test('then the client-side files are created', () {
-      expect(
-        codeMap.keys,
-        contains(path.join(
-          '..',
-          'example_project_client',
-          'lib',
-          'src',
-          'protocol',
-          'example.dart',
-        )),
-        reason: 'Expected client-side file to be present, found none.',
-      );
-
-      expect(
-        codeMap.keys,
-        contains(path.join(
-          '..',
-          'example_project_client',
-          'lib',
-          'src',
-          'protocol',
-          'user.dart',
-        )),
-        reason: 'Expected client-side file to be present, found none.',
-      );
-    });
   });
 
   test(
-      'Given a server-side only class when generating the code then the client-side file is NOT created',
+      'Given a server-side only class when generating the code then the server-side file is created',
       () {
     var entities = [
       ClassDefinitionBuilder()
@@ -141,17 +82,8 @@ void main() {
 
     expect(
       codeMap.keys,
-      isNot(
-        contains(path.join(
-          '..',
-          'example_project_client',
-          'lib',
-          'src',
-          'protocol',
-          'example.dart',
-        )),
-      ),
-      reason: 'Expected client-side file to NOT be present, found one.',
+      contains(path.join('lib', 'src', 'generated', 'example.dart')),
+      reason: 'Expected server-side file to be present, found none.',
     );
   });
 }
